TA5653 Extract common pipeline library
Introduced smartlogic-common pipeline library with common code to be used
in individual projects Jenkinsfile's.

The library is configured as a global trusted library in Jenkins under
'smartlogic-common' name and default version of 'master'.

The library exposes the 'smartlogic' global variable which currently supports
the following methods:
- buildDockerImage
- buildInDocker
- buildNpmPackage

All types of builds support company-wide features such as:
* timestamping/coloring of console out
* ability to merge any branch to master
* send email/flowdock notifications
* docker-compatible way of cleaning workspace

Each Jenkinsfile of our projects has been simplified by using commonized methods.

The merge-to-master functionality has been refactored so that it is performed
in separate dedicated docker image "smartlogic/merge-tool". This way, it does
not depend on the image where main build is performed, and also supports builds
done outside of docker.

Implemented by pp
Reviewed by mwo